refactor(repository): move common logic in a base class for sync/async strawchemy repositories
